+++ src/main/java/com/caffeine/gwanghwamun/controller/TestController.java
@@ -1,5 +1,7 @@
-package com.caffeine.gwanghwamun;
+package com.caffeine.gwanghwamun.controller;
 
+import com.caffeine.gwanghwamun.domain.ExampleEntity;
+import com.caffeine.gwanghwamun.repository.TestRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
