+++ src/main/java/com/caffeine/gwanghwamun/domain/ExampleEntity.java
@@ -1,4 +1,4 @@
-package com.caffeine.gwanghwamun;
+package com.caffeine.gwanghwamun.domain;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
